+++ packages/verify-contract/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/verify-contract
 
+## 1.5.2
+
+### Patch Changes
+
+- 128b697: fix duplicated sourcemaps
+
 ## 1.5.1
 
 ### Patch Changes